+++ crates/operator_impl/src/incredible_square_operator.rs
