+++ crates/operator_impl/src/lib.rs
@@ -1,3 +1,3 @@
 //! Operator with alloy rpc client to send signed task response to aggregator
 /// Operator builder
-pub mod builder;
+pub mod incredible_square_operator;
